refactor(number-input): ♻️  remove focused & refactor
