+++ src/number-input/helpers/__keys.ts
@@ -1,0 +1,2 @@
+// Automatically generated
+export const USE_SPIN_BUTTON_KEYS = [] as const;
